% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cr_level.R
\name{add_cr_level}
\alias{add_cr_level}
\title{Add convergent recombination level to each clone record}
\usage{
add_cr_level(.data, .clonal_seq, ...)
}
\arguments{
\item{.data}{data frame which of the TCR repertoire's clone records. Each
row of the resulting data frame contains information about a single TCR
component sequence in one cell of a particular CDR3 sequence.}

\item{.clonal_seq}{Column variable name of the CDR3's nucleotide sequence.}

\item{...}{additional grouping column variables that might be necessary to
distinguish merged data frames of multiple repertoires, or clonal sequence
specification (such as V/J gene segments).}
}
\value{
data frame with \code{CR_level} column
}
\description{
Calculate unique number of nucleotide sequences encoding a specific amino
acid sequence. This value associated with TCR repertoire clonal sequences
and termed as "CR-level" (convergent recombination level). (CR-level)
}
\details{
several clones that encoded the same amino acid sequence were found to be
structurally distinct at the nucleotide level, strongly implying clonal
selection and expansion is operating at the level of specific TCR-peptide
interactions.
}
\examples{

df <- data.frame(
  group_id=gl(2,4,8,labels = c('cancer','control')),
  sample_id=gl(4,2,8,labels = LETTERS[1:4]),
  nt=c('CGCGTGAAG', 'CGGGTGAAG','CACGAA','AAGGGGTCCGTG',
       'AAGGGGTCCGTC','CGGGTGAAG','AAGGGGTCCGTT','CGGGTGAAG'),
  aa=c('RVK','RVK','HE','KGSV','KGSV','RVK','KGSV','RVK')

)

add_cr_level(df, nt)
add_cr_level(df, nt, sample_id)

require('dplyr')

group_by(df, group_id) \%>\% add_cr_level(nt)
group_by(df, group_id) \%>\% add_cr_level(nt, sample_id)

}
