% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_share_level.R
\name{add_share_level}
\alias{add_share_level}
\title{Add share level of clonotypes among repertoires}
\usage{
add_share_level(.data, .clonotype, .rep_id, ...)
}
\arguments{
\item{.data}{repertoire data frame}

\item{.clonotype}{clonotype sequence variable}

\item{.rep_id}{repertoire identification column}

\item{...}{additional variables to group by}
}
\value{
data frame with 'share' column
}
\description{
Add share level of clonotypes among repertoires
}
\details{
Previous studies have shown that the extent of sharing and
the clonotypic frequency of TCRb sequences are significantly
correlated with their production efficiencies in simulations
of a random recombination process because of the phenomenon
of convergent recombination.
https://doi.org/10.1073/pnas.1319389111 (REMOVE) ###########################################
}
\examples{

df <- data.frame(
  group_id=gl(2,4,8,labels = c('cancer','control')),
  sample_id=gl(4,2,8,labels = LETTERS[1:4]),
  nt=c('CGCGTGAAG', 'CGGGTGAAG','CACGAA','AAGGGGTCCGTG',
       'AAGGGGTCCGTC','CGGGTGAAG','AAGGGGTCCGTT','CGGGTGAAG'),
  aa=c('RVK','RVK','HE','KGSV','KGSV','RVK','KGSV','RVK')
)

add_share_level(df, aa, sample_id)

require('dplyr')

group_by(df, group_id) \%>\% add_share_level(aa, sample_id)

}
