% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_clonotype_diversity.R
\name{calculate_clonotype_diversity}
\alias{calculate_clonotype_diversity}
\title{Calculate the number of unique sequences coding for a given clonotype in a TCR repertoire}
\usage{
calculate_clonotype_diversity(repertoire, clonotype, clone_col, clonotype_col)
}
\arguments{
\item{repertoire}{a data frame containing the TCR repertoire, with columns for the clonotype and the sequence}

\item{clonotype}{a character string specifying the clonotype to calculate the diversity for}
}
\value{
an integer representing the number of unique sequences coding for the specified clonotype
}
\description{
Calculate the number of unique sequences coding for a given clonotype in a TCR repertoire
}
