% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_clonotype_frequency.R
\name{calculate_clonotype_frequency}
\alias{calculate_clonotype_frequency}
\title{Calculate the fraction of the total number of sequences in a TCR repertoire that are represented by a given clonotype}
\usage{
calculate_clonotype_frequency(repertoire, clonotype, clone_col, clonotype_col)
}
\arguments{
\item{repertoire}{a data frame containing the TCR repertoire, with columns for the clonotype and the sequence}

\item{clonotype}{a character string specifying the clonotype to calculate the frequency for}
}
\value{
a numeric value representing the fraction of the total number of sequences in the repertoire that are represented by the specified clonotype
}
\description{
Calculate the fraction of the total number of sequences in a TCR repertoire that are represented by a given clonotype
}
