% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_shared_clonotypes.R
\name{calculate_shared_clonotypes}
\alias{calculate_shared_clonotypes}
\title{Calculate the level of sharing of a given clonotype among different TCR repertoires}
\usage{
calculate_shared_clonotypes(repertoire_list, clonotype)
}
\arguments{
\item{repertoire_list}{a list of data frames, each containing a TCR repertoire with columns for the clonotype and the sequence}

\item{clonotype}{a character string specifying the clonotype to calculate the level of sharing for}
}
\value{
a numeric value representing the level of sharing of the specified clonotype among the different TCR repertoires
}
\description{
Calculate the level of sharing of a given clonotype among different TCR repertoires
}
