% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_prepare.R
\name{clonalSeq2DF}
\alias{clonalSeq2DF}
\title{repertoire's clonal sequence to data frame with clonotype sequence}
\usage{
clonalSeq2DF(clonal_seq, clone_var)
}
\arguments{
\item{clonal_seq}{character vector / data frame}

\item{clone_var}{(optional) for data frame with more than one column, provide the CDR3's clonal sequence (nucleation sequence) column name}
}
\value{
data frame
}
\description{
repertoire's clonal sequence to data frame with clonotype sequence
}
\examples{

clonal_seq <- c("ATG", "TTC", "TTC", "TTT", "ATG", "ATG")

clonalSeq2DF(clonal_seq)

df <- data.frame(cdr3_nt=clonal_seq)

clonalSeq2DF(df)


}
