% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr.R
\name{clone_gen}
\alias{clone_gen}
\title{generate sequence with coding codons}
\usage{
clone_gen()
}
\value{
string
}
\description{
generate sequence with coding codons
}
\examples{

nt_gen <- clone_gen()
nt_gen(3)
}
