% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{convert_sequence}
\alias{convert_sequence}
\title{Convert a vector of nucleotide sequences into a vector of amino acid sequences.}
\usage{
convert_sequence(nt_vec)
}
\arguments{
\item{nt_vec}{A vector of nucleotide sequences. Each sequence must contain consecutive
triplets of 'A','G','T','C' (uppercase)}
}
\value{
A vector of amino acid sequences.
}
\description{
Convert a vector of nucleotide sequences into a vector of amino acid sequences.
}
\examples{
translate(c("ATGAGACCCAGG", "ATGAGACCCAGG"))
  #> c("MGR", "MGR")

}
