% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_class.R
\name{cr_class}
\alias{cr_class}
\title{Title x}
\usage{
cr_class(.gruop, .sub_gruop, ..., .public_class = 2, .exclusive_class = 1)
}
\arguments{
\item{.gruop}{x}

\item{.sub_gruop}{x}

\item{...}{x}

\item{.public_class}{x}

\item{.exclusive_class}{x}
}
\value{
x
}
\description{
Title x
}
\examples{


require(dplyr)

df <- data.frame(x=factor(LETTERS[sample(1:5, 5,replace = TRUE)]),
                 y=factor(letters[sample(1:2, 5, replace = TRUE)]),
                 clonotype=sample(c("ASD", "AWD"), 5, replace = TRUE, prob = c(.2,.8)))


df \%>\%
  group_by(clonotype) \%>\%
  summarise(cr=cr_class(.gruop = y, .sub_gruop = x))

}
