% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_class.R
\name{cr_class_df}
\alias{cr_class_df}
\title{Title x}
\usage{
cr_class_df(
  rep_groups,
  clonotype,
  rid,
  gid,
  ...,
  min_shared = 1,
  min_public = 2,
  exclusive_max = 1
)
}
\arguments{
\item{rep_groups}{x}

\item{clonotype}{x}

\item{rid}{x}

\item{gid}{x}

\item{...}{x}

\item{min_shared}{x}

\item{min_public}{x}

\item{exclusive_max}{c}
}
\value{
x
}
\description{
Title x
}
\examples{

p <- replicate(4, rand_group(), FALSE)

df <- groupList2DF(p)

df <- cbind(df, pid=ifelse(as.numeric(df$gid)\%\%2, "human", "mouse"))

head(df)

cr_df <- cr_class_df(df, clonotype, rid, gid, pid)
head(cr_df)

}
