% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_class.R
\name{cr_class_tbl}
\alias{cr_class_tbl}
\title{Title x}
\usage{
cr_class_tbl(
  rep_groups,
  clonotype,
  rid,
  gid,
  ...,
  min_freq = 1,
  public_min = 2,
  exclusive_max = 1
)
}
\arguments{
\item{rep_groups}{x}

\item{clonotype}{x}

\item{rid}{x}

\item{gid}{x}

\item{...}{x}

\item{min_freq}{x}

\item{public_min}{x}

\item{exclusive_max}{x}
}
\value{
x
}
\description{
Title x
}
\examples{

groupList2DF(l) \%>\% dplyr::rename(group="gid") \%>\% dplyr::mutate(pid=gl(2,1,nrow(.)))  \%>\% cr_class_df(clonotype, rid, group, pid)


}
