% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_clonetype.R
\name{cr_clonotype}
\alias{cr_clonotype}
\title{Title}
\usage{
cr_clonotype(v1, v2, ...)
}
\arguments{
\item{v1}{character vector of the first populations}

\item{v2}{character vector of the second populations}

\item{...}{any additional character vectors}
}
\value{
list of private, inclusive and exclusive clonotype
}
\description{
Title
}
