% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_freq}
\alias{cr_freq}
\title{Title x}
\usage{
cr_freq(.freq, .clone, .apply_freq = sum, .apply_clone = mean)
}
\arguments{
\item{.freq}{x}

\item{.clone}{z}

\item{.apply_freq}{x}

\item{.apply_clone}{x}
}
\value{
x
}
\description{
Title x
}
\examples{

require(dplyr)

df <- data.frame(x=factor(LETTERS[sample(1:5, 10,replace = TRUE)]),
                 y=factor(letters[sample(1:2, 10, replace = TRUE)]),
                 clonotype=sample(c("ASD", "AWD"), 10, replace = TRUE),
                 Freq=runif(10))

df \%>\%

  group_by(clonotype, y) \%>\%

  summarise(cr_lvl=cr_freq(Freq, x))

}
