% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level}
\alias{cr_level}
\title{CR level table}
\usage{
cr_level(...)
}
\arguments{
\item{...}{repertoire list / data frames vectors}
}
\value{
table
}
\description{
CR level table
}
\examples{

# generate random
g <- replicate(4, rand_group(), simplify = FALSE)

cr_level(g)

}
