% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level_clonotype}
\alias{cr_level_clonotype}
\title{CR-level of a selected clonotypes}
\usage{
cr_level_clonotype(rep, of, clonotype_col, clone_col)
}
\arguments{
\item{rep}{df}

\item{of}{sequences}

\item{clonotype_col}{name}

\item{clone_col}{name}
}
\description{
CR-level of a selected clonotypes
}
\examples{

rand_rep_df(seq_n=1000, seq_l=3) \%>\%

  cr_level_clonotype(c("S", "T", "M"), "aa","nt")

}
