% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level_df}
\alias{cr_level_df}
\title{Convergent Recombination level for data frame's clonal sequence column}
\usage{
cr_level_df(df, ..., clone_var, clonotype_var)
}
\arguments{
\item{df}{data frame}

\item{...}{Column names to group the clones by as distinct repertoires}

\item{clone_var}{Column name of CDR3's nucleotide sequence}

\item{clonotype_var}{(optional) Column name of CDR3's amino acid sequence}
}
\value{
data frame
}
\description{
calculate unique number of nucleotide sequences encoding a specific amino
acid sequence. This value associated with TCR repertoire clonal sequences
and termed as "CR-level" (convergent recombination level).
}
\examples{

df <- data.frame(
  group_id=gl(2,4,8,labels = c('cancer','control')),
  sample_id=gl(4,2,8,labels = LETTERS[1:4]),
  nt=c('CGGGTGAAG', 'CGGGTGAAG','CACGAA','AAGGGGTCCGTG',
       'AAGGGGTCCGTC','CGGGTGAAG','AAGGGGTCCGTT','CGGGTCAAG'),
  aa=c('RVK','RVK','HE','KGSV','KGSV','RVK','KGSV','RVK')
)

cr_level_df(df, clone_var=nt, clonotype_var=aa)

cr_level_df(df, clone_var=nt)

cr_level_df(df, group_id, sample_id, clone_var=nt, clonotype_var=aa)

}
