% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level_group}
\alias{cr_level_group}
\title{Make a table that specifies the CR values for each individual's clonotypes.}
\usage{
cr_level_group(clonotype_list, clonotype_col, clone_col)
}
\arguments{
\item{clonotype_list}{}
}
\value{
return a matrix with the cr-level of the individual in the column
in each row of the clonotype
}
\description{
claculate the CR level in each clonotype in every sample.
The number of various NT sequences encoding a particular AA sequence
is defined as the CR level in this context.
}
\examples{
# From the letter vector representing the clonotype sequences,
# generate a random list of letters.

rep_list <- rand_group(seq_n = 3)

cr_level_group(rep_list, "aa", "nt")

dfl <- rand_group(seq_n = 3) \%>\% lapply(function(df) df[,2:1]) \%>\%

   lapply(setNames, names(formals(cr_level)))



}
