% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level_matrix}
\alias{cr_level_matrix}
\title{Make a table that specifies the CR values for each individual's clonotypes.}
\usage{
cr_level_matrix(clonotype_list)
}
\arguments{
\item{clonotype_list}{}
}
\value{
return a matrix with the cr-level of the individual in the column
in each row of the clonotype
}
\description{
claculate the CR level in each clonotype in every sample.
The number of various NT sequences encoding a particular AA sequence
is defined as the CR level in this context.
}
\examples{

# From the letter vector representing the clonotype sequences,
# generate a random list of letters.
l <- replicate(6, sample(LETTERS, sample(13:18, 1, replace = TRUE)), simplify = FALSE)

cr_level_matrix(l)

}
