% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level_tbl}
\alias{cr_level_tbl}
\title{CR-level throughout all sub-groups matrix}
\usage{
cr_level_tbl(df, ..., clonal_var, clonotype_var)
}
\arguments{
\item{df}{data frame}

\item{...}{repertoire ID and any other group ID column}

\item{clonotype_var}{clonotype seq variable name}

\item{clone_var}{clonal seq variable name}
}
\value{
matrix
}
\description{
CR-level throughout all sub-groups matrix
}
\examples{


df <- data.frame(nt=c("ATG","TTC","TAT","TTT","ATG","ATG"),
                 aa=c("M","F","Y","F","M","M"),
                 rep_id = gl(3,1, 6),
                 group_id = gl(2, 3, labels = c("cancer", "control")))

cr_level_tbl(df, clone_var = nt, clonotype_var = aa, group_id, rep_id)

}
