% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R, R/cr_number.R
\name{cr_number}
\alias{cr_number}
\title{Calculate convergent recombination level by the number of unique nucleotide sequences}
\usage{
cr_number(nucleotide, ...)

cr_number(nucleotide, ...)
}
\arguments{
\item{nucleotide}{A character vector of DNA sequences representing the TCR sequence to be analyzed}

\item{...}{Additional arguments (currently not used)}
}
\value{
A numeric value indicating the convergent recombination level of the TCR sequence

A numeric value indicating the convergent recombination level of the TCR sequence
}
\description{
This function calculates the convergent recombination level of a TCR sequence by counting the number of unique nucleotide sequences in the sequence. The TCR sequence is given as a character vector of DNA sequences in the \code{.clone} argument.

This function calculates the convergent recombination level of a TCR sequence by counting the number of unique nucleotide sequences in the sequence. The TCR sequence is given as a character vector of DNA sequences in the \code{.clone} argument.
}
