% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_source_plot.R
\name{cr_source}
\alias{cr_source}
\title{Title}
\usage{
cr_source(input_list)
}
\arguments{
\item{input_list}{}
}
\description{
Title
}
\examples{

g <- list(rep1=rand_rep_vec(seq_type = "nt", 7, rep(5, 7)),
          rep2=rand_rep_vec(seq_type = "nt", 9, rep(5, 9)))

}
