% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_subgroup.R
\name{cr_subgroup}
\alias{cr_subgroup}
\title{CR-level for multiple repertoires}
\usage{
cr_subgroup(.clone, .id, .subgroup, .subgroup_func = mean)
}
\arguments{
\item{.subgroup}{}
}
\description{
CR-level for multiple repertoires
}
