% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_clonetype.R
\name{cr_type}
\alias{cr_type}
\title{Title}
\usage{
cr_type(count_table)
}
\arguments{
\item{count_table}{column is the number of samples having a specific clonotype in a row}
}
\value{
list of clonotype in each sub population
}
\description{
Title
}
