% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crtcr-package.R
\docType{package}
\name{crtcr-package}
\alias{crtcr}
\alias{crtcr-package}
\title{crtcr: Convergent recombination of TCR repertoire}
\description{
Functions for computations relating to convergent recombination for TCR repertoire sequencesing data are included in this package. On the basis of these supplied calculations, identical clonotypes created from numerous clonal sequences and belonging exclusively to a particular immunological state are being obtained and could be used for subsequent analysis approach.
}
\author{
\strong{Maintainer}: Ido Hasson \email{idoh@systemsbiomed.org} (\href{https://orcid.org/0000-0001-8382-5638}{ORCID})

Authors:
\itemize{
  \item Sol Efroni
  \item Hagit Philip
  \item Alona Zilberberg
}

}
\keyword{internal}
