% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_class.R
\name{cr_freq_class}
\alias{cr_freq_class}
\title{function that takes a vector of frequencies and returns a character vector of class labels.}
\usage{
cr_freq_class(c_freq, min_freq = 1, public_min = 2, exclusive_max = 1)
}
\arguments{
\item{c_freq}{numerical vector}

\item{min_freq}{minimum numerical value}

\item{public_min}{minimum sum from all frequencies having a minimal value to be considered as public clonotype}

\item{exclusive_max}{maximal number of public clonotypes groups to be considered as public exclusive. o.w. inclusive.}
}
\value{
character
}
\description{
function that takes a vector of frequencies and returns a character vector of class labels.
}
\section{Functions}{
\itemize{
\item \code{cr_freq_class()}: clonotype frequencies in each group, calculate its convergent
recombination public classifications based on various conditions.

}}
\examples{

cr_freq_class(c(1,0,1,0)) # public inclusive
cr_freq_class(c(2,0,0,0)) # public exclusive
cr_freq_class(c(1,0,0,0)) # private
cr_freq_class(c(0,0,0,0)) # NA if no clonotype exists
cr_freq_class(c(1,2,0,10), min_freq=2) # public inclusive
cr_freq_class(c(1,2,0,10), min_freq=3) # public exclusive
cr_freq_class(c(1,2,0,10), min_freq=11) # also NA if non frequencies ai sufficient
cr_freq_class(c(1,2,0,10), min_freq=2, exclusive_max=2) # public exclusive
cr_freq_class(c(1,2,0,10), min_freq=1, exclusive_max=2) # public inclusive
cr_freq_class(c(0, 1, 2, 3), min_freq=1, public_min=5) # public inclusive
cr_freq_class(c(0, 1, 2, 4), min_freq=2, public_min=5) # public exclusive
cr_freq_class(c(0, 1, 2, 4), min_freq=3, public_min=5) # private



}
