% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_join.R
\name{group_join}
\alias{group_join}
\title{Join all repertoires' clonotypes data into one data frame}
\usage{
group_join(...)
}
\arguments{
\item{...}{repertoire group}
}
\value{
data frame
}
\description{
Join all repertoires' clonotypes data into one data frame
}
\examples{

# generate random
g <- replicate(4, rand_group(), simplify = FALSE)

group_join(g)

}
