% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_count.R
\name{group_value_counts}
\alias{group_value_counts}
\title{Calculates the number of unique values in each group of a given vector}
\usage{
group_value_counts(
  value_vector,
  grouping_vector,
  proportion = FALSE,
  remove_na = FALSE
)
}
\arguments{
\item{value_vector}{A vector of nucleotide or amino acid values}

\item{grouping_vector}{A vector indicating which group each element of the 'value_vector' belongs to}

\item{proportion}{A logical indicating whether to return the proportion of unique values}

\item{remove_na}{A logical indicating whether to exclude NA values from the calculation}
}
\value{
A vector of the number of unique values in each group of the input vector, or the proportion of unique values
in each group if the 'proportion' argument is set to TRUE
}
\description{
This function splits a vector of values into groups based on a grouping vector, and then calculates the number
of unique values in each group using the 'unique_value_count' function. The type of values
(nucleotide or amino acid) is automatically detected based on the characters in the vector.
The calculation can be adjusted to exclude NA values and return the proportion of unique values instead of the count.
}
