% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_prepare.R
\name{groupList2DF}
\alias{groupList2DF}
\title{join groups of repertoire list into a dataframe}
\usage{
groupList2DF(group_list, ...)
}
\arguments{
\item{group_list}{list-of-list of character vector or data frames}

\item{...}{for data frames with more than one column, clonal sequence column name is needed. adding also clonotype column name is suggested}
}
\value{
data frame
}
\description{
join groups of repertoire list into a dataframe
}
\examples{

vec_list1 <- list(
A=c("ATG", "TTC", "TTT"),
B=c("TTT", "ATG", "ATG")
)
vec_list2 <- list(
  C=c("ATG", "TTC", "TTT"),
  D=c("TTT", "ATG", "ATG")
)

grouped_rep_list <- list(group1=vec_list1,group2=vec_list2)

groupList2DF(grouped_rep_list)

}
