% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_exclusive.R
\name{is_exclusive}
\alias{is_exclusive}
\title{Is exclusive public clonotype}
\usage{
is_exclusive(.subgroup, ..., .exclusive = 1)
}
\arguments{
\item{.subgroup}{sub-group identification vector.}

\item{...}{additional identifier vectors}

\item{.exclusive}{maximal number of subgroups a clone is found in.}
}
\value{
TRUE for exclusive, FALSE for inclusive.
}
\description{
True for if clonotype’s present across subgroups doesn't surpass the threshold
}
\examples{

is_exclusive(c("A", "A", "A", "A"))
is_exclusive(c("A", "A", "A", "A", "B"))

}
