% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_logic.R
\name{is_public}
\alias{is_public}
\title{logical check if a frequencies vector is considered public with optional modifications}
\usage{
is_public(freq, min_freq = 1L, min_rep = 2L)
}
\arguments{
\item{freq}{numerical vector}

\item{min_freq}{minimal value to be counted as repertoire having a clonotype}

\item{min_rep}{minimal number of repertoire having a clonotype to be conciser public}
}
\value{
TRUE if its public, FALSE if private, and NA if none of the frequencies are sufficient
}
\description{
logical check if a frequencies vector is considered public with optional modifications
}
\examples{

is_public(c(1, 0 , 2, 4), min_freq=3)

}
