% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_public_freq.R
\name{is_public_freq}
\alias{is_public_freq}
\title{Is public clonotype by frequencies}
\usage{
is_public_freq(.freq, .public = 2, na.rm = FALSE)
}
\arguments{
\item{.freq}{numerical vector}

\item{.public}{value count threshold}

\item{na.rm}{if FALSE (the default), return NA when all frequency values are zero.}
}
\value{
TRUE for public, False for private clonotype
}
\description{
Returns True if clonotype's number of frequency values which are non-zero throughout repertoires / group, meet with the condition. if so, the clonotype is public
}
\examples{

is_public_freq(c(0,0,1)) # FALSE
is_public_freq(c(1,2,0)) # TRUE

is_public_freq(c(0,0,0)) # NA
is_public_freq(c(0,0,0), na.rm = TRUE) # FALSE

is_public_freq(c(0,0,1), .5) # FALSE
is_public_freq(c(1,2,0), .5) # TRUE
is_public_freq(c(1,2,0), .75) # FALSE

}
