% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr.R
\name{nt2df}
\alias{nt2df}
\title{Pair NT vector to AA to a new DF}
\usage{
nt2df(nt)
}
\arguments{
\item{nt}{char vector}
}
\value{
df
}
\description{
Pair NT vector to AA to a new DF
}
\examples{

nt2df(c("AGT", "ATT"))

}
