% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_class.R
\name{public_class}
\alias{public_class}
\title{Public classification}
\usage{
public_class(.id, ..., .subgroup = NULL, .public = 2, .exclusive = 1)
}
\arguments{
\item{.id}{repertoire identifier vector}

\item{...}{additional identifier vectors.}

\item{.subgroup}{sub-group identification vector. for the default value, NULL, returns "public" instead of "exclusive” / “inclusive”}

\item{.public}{minimum number of repertoires.}

\item{.exclusive}{maximal number of subgroups a clone is found in.}
}
\value{
"private" / "exclusive” / “inclusive”
}
\description{
returns the publicness classifications "private," "exclusive,” and
“inclusive” by the repertoires’ presence, throughout all subgroups.
}
\examples{

public_class(c("A", "B"), .subgroup=c("C", "D"))
public_class(c("A", "B"), .subgroup=c("C", "C"))
public_class(c("A", "A"), .subgroup=c("C", "C"))

public_class(c("A", "B"))
public_class(c("A", "A"))

}
