% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_class_freq.R
\name{public_class_freq}
\alias{public_class_freq}
\title{Public classification of a clonotype}
\usage{
public_class_freq(
  .freq,
  .subgroup = NULL,
  .public = 2,
  .exclusive = 1,
  na.rm = FALSE
)
}
\arguments{
\item{.freq}{numerical vector}

\item{.subgroup}{identifier vector (optional)}

\item{.public}{value count threshold}

\item{.exclusive}{maximal number of subgroups a clone is found in.}

\item{na.rm}{if FALSE (the default), return NA when all frequency values are zero.}
}
\value{
character value ("private" / "public" / "exclusive” / “inclusive”)
}
\description{
Public classification of a clonotype
}
\examples{

public_class_freq(c(1,1,0,0,0), c("A", "A", "B", "B", "C"))

public_class_freq(c(1,1,0,0,0), c("A", "A", "B", "B", "C"), .public=.5)

public_class_freq(c(1,1,0,0,1), c("A", "A", "B", "B", "C"), .exclusive=2)

}
