% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_levels.R
\name{quantile_levels}
\alias{quantile_levels}
\title{Compute levels of a numeric vector based on proportions and quantiles}
\usage{
quantile_levels(values, top = c(0.25, 0.5, 0.75), rm.lvl = TRUE)
}
\arguments{
\item{values}{a numeric vector}

\item{top}{a numeric vector of quantiles to use when computing the levels}

\item{rm.dup}{a logical value indicating whether duplicate levels should be removed}
}
\value{
a numeric vector with the computed levels
}
\description{
This function computes the levels of a numeric vector based on the proportions of values in the vector and a set of quantiles provided as input. It uses the \code{table}, \code{proportions}, and \code{cumsum} functions to calculate the proportions of values in the vector and the corresponding cumulative proportions, and then uses the \code{vapply} function to apply a function to the provided quantiles to find the first levels that are higher than the quantiles. Finally, it returns the resulting levels, optionally removing any duplicates.
}
\examples{
# Compute the levels for the numbers 1, 2, 3, 4, and 5 using the quantiles 0.25, 0.5, and 0.75
top_prec(c(1, 2, 3, 4, 5), c(0.25, 0.5, 0.75))

# Compute the levels for the numbers 1, 2, 3, 4, and 5 using the quantiles 0.25, 0.5, and 0.75, and remove duplicates
top_prec(c(1, 2, 3, 4, 5), c(0.25, 0.5, 0.75), rm.dup = TRUE)
}
