% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_group.R
\name{rand_group}
\alias{rand_group}
\title{random list of NT vector sequnces}
\usage{
rand_group(rep_n = rpois(1, 10))
}
\arguments{
\item{rep_n}{number of vector}
}
\value{
list of vector
}
\description{
random list of NT vector sequnces
}
\examples{

# also random group size for no rep_n arg
rand_group()

}
