% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_prepare.R
\name{rep2DF}
\alias{rep2DF}
\title{repertoire to data frame with clonal sequence and clonotype sequence}
\usage{
rep2DF(rep, clone_var, clonotype_var)
}
\arguments{
\item{rep}{character vector or dataframe}

\item{clone_var}{column name of the clonal sequences}

\item{clonotype_var}{(optional) column name of the clonotype sequences}
}
\value{
dataframe
}
\description{
repertoire to data frame with clonal sequence and clonotype sequence
}
\examples{

clonal_seq <- c("ATG", "TTC", "TTC", "TTT", "ATG", "ATG")
rep2DF(clonal_seq)

df1 <- data.frame(cdr3_nt=clonal_seq)
rep2DF(df1)
rep2DF(df1, clone_var=cdr3_nt)

clonotype_seq <- c("M", "F", "F", "F", "M", "M")

df2 <- data.frame(cdr3_nt=clonal_seq, cdr3_aa=clonotype_seq)
rep2DF(df2, clone_var=cdr3_nt, clonotype_var=cdr3_aa)
rep2DF(df2, clone_var=cdr3_nt) # translate 'cdr3_nt' to amino acid sequences


}
