% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_join.R
\name{replist2DF}
\alias{replist2DF}
\title{convert list of nucleation sequences to data frame}
\usage{
replist2DF(rep_list, rep_names)
}
\arguments{
\item{rep_list}{list}

\item{rep_names}{sample names}
}
\value{
data frame
}
\description{
convert list of nucleation sequences to data frame
}
\examples{

nt_list <- rand_group()

group_df <- replist2DF(nt_list, rep_names=LETTERS[seq_along(nt_list)])

head(group_df)

}
