% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_level.R
\name{share.level}
\alias{share.level}
\title{Title}
\usage{
share.level(clone_list, method = "unique")
}
\arguments{
\item{clone_list}{x}

\item{method}{x}
}
\value{
x
}
\description{
x
}
\examples{

rep_list <- rand_group()

share.level(rep_list, method="relative")

}
