% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_level.R
\name{share_level}
\alias{share_level}
\title{Share level}
\usage{
share_level(.r_id, ..., share_func = NULL, na.rm = FALSE)
}
\arguments{
\item{...}{x}

\item{share_func}{x}

\item{na.rm}{x}
}
\value{
x
}
\description{
Share level
}
\examples{

df <- data.frame(
  x=c("A", "A", "A", "B", "B", "B"),
  y=c("C", "C", "D", "D", "E", "E"),
  clonotype=c("a", "b", "a", "b", "a", "b")
)

share_level(df)

require(dplyr)

my_func <- function(df) {

 table(df) \%>\%

   rowMeans() \%>\%

   sum()
}

group_by(df, clonotype, x) \%>\%
  summarise(share=share_level(x,y, share_func = my_func))

}
