% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_level.R
\name{share_level_df}
\alias{share_level_df}
\title{Compute the share level of clonotypes in a group}
\usage{
share_level_df(data, clonotype_var, rid, ...)
}
\arguments{
\item{data}{data frame}

\item{clonotype_var}{clonotype sequence column name}

\item{rid}{repertoire identification column name}

\item{...}{group by}
}
\value{
data frame
}
\description{
Compute the share level of clonotypes in a group
}
\examples{

require(dplyr)

df <- data.frame(nt=c("ATG","TTC","TAT","TTT","ATG","ATG"),
                 aa=c("M","F","Y","F","M","M"),
                 rep_id = gl(3,1, 6),
                 group_id = gl(2, 3, labels = c("cancer", "control")))
df \%>\%
group_by(group_id) \%>\%
share_level_df(aa, rep_id)

# Same as:
share_level_df(df, aa, rep_id, group_id)

}
