% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_level.R
\name{share_level_table}
\alias{share_level_table}
\title{Title x}
\usage{
share_level_table(df, clonotype_var, rid, gid, ...)
}
\arguments{
\item{df}{x}

\item{clonotype_var}{x}

\item{rid}{x}

\item{gid}{x}

\item{...}{x}
}
\value{
x
}
\description{
Title x
}
\examples{

df <- data.frame(nt=c("ATG","TTC","TAT","TTT","ATG","ATG"),
                 aa=c("M","F","Y","F","M","M"),
                 rep_id = gl(3,1, 6),
                 group_id = gl(2, 3, labels = c("cancer", "control")))

 share_level_table(df, aa, rep_id, group_id)

}
