% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_table.R
\name{share_table}
\alias{share_table}
\title{Title}
\usage{
share_table(df, clonotype, group_id, rep_id, ...)
}
\arguments{
\item{df}{data frame with clonotype column, group ID coliumn and repertoire ID column}

\item{clonotype}{clonotype column name}

\item{group_id}{Group ID column name}

\item{rep_id}{Repertoire ID column name}

\item{...}{any additional column name}
}
\value{
table
}
\description{
Title
}
\examples{

df_lists <- replicate(4, rand_group(), FALSE)
rand_df <- group_join(df_lists)
share_table(rand_df, "clonotype", "group", "rep_id")

}
