% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharing.R
\name{sharing}
\alias{sharing}
\title{Title}
\usage{
sharing(..., by = c("clonotype", "rep_id"))
}
\arguments{
\item{...}{data frames or vectors}

\item{by}{x}
}
\value{
table
}
\description{
Title
}
\examples{

nt_gen <- clone_gen()

group1 <- list(sapply(rpois(100, 2)+1, nt_gen), sapply(rpois(100, 2)+1, nt_gen))
group2 <- list(sapply(rpois(100, 2)+1, nt_gen), sapply(rpois(100, 2)+1, nt_gen))

group_join(list(a=group1, b=group2))#'

}
