% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharing_number.R
\name{sharing_number}
\alias{sharing_number}
\title{Number of distinct repertoires}
\usage{
sharing_number(.id, ...)
}
\arguments{
\item{.id}{vector of repertoire's identifier}

\item{...}{additional identifiers to distinguish between same \code{.id} value.}
}
\value{
integer of the repertoire count.
}
\description{
Number of distinct repertoires
}
\examples{

v1 <- c("A", "A", "A", "B", "A", "B")
sharing_number(v1)

(v2 <- rep(1:2, each=3))
sharing_number(v1, v2)

(v3 <- rep(1:2, times=3))
sharing_number(v1, v2, v3)

}
