% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_clonality.R
\name{summarize_clonality}
\alias{summarize_clonality}
\title{Summarize the clonality of a TCR repertoire}
\usage{
summarize_clonality(repertoire, clonotype, sequence)
}
\arguments{
\item{repertoire}{a data frame containing the TCR repertoire, with columns for the clonotype and the sequence}

\item{sequence}{}
}
\value{
a data frame containing clonality measures (e.g., number of unique sequences, frequency, and level of sharing) for each clonotype in the repertoire
}
\description{
Summarize the clonality of a TCR repertoire
}
