% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_join.R
\name{translate}
\alias{translate}
\title{translate the NT sequence to AA sequence}
\usage{
translate(nt_vec)
}
\arguments{
\item{nt_vec}{a character vector}
}
\value{
a character vector
}
\description{
translate the NT sequence to AA sequence
}
\examples{

translate("ATG")

}
