% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_n.R
\name{unique_n}
\alias{unique_n}
\title{Amount of unique values overall.}
\usage{
unique_n(..., na.rm = FALSE)
}
\arguments{
\item{...}{Same length vectors of values.}

\item{na.rm}{if TRUE, ignore aligned values if any of them are NA. For FALSE,
the default, NA is treated as an unique value.}
}
\value{
An integer of unique value count.
}
\description{
Amount of unique values overall.
}
\examples{

v1 <- c("A", "A", "A", "B")
v2 <- factor(c("seqA", "seqA", "seqB", "seqB"))
v3 <- c(1,NA,2,2)

unique_n(v1)
unique_n(v1, v2)
unique_n(v1, v2, v3) # NA in v3 is considered as a unique value.
unique_n(v1, v2, v3, na.rm = TRUE)

}
