% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_count.R
\name{unique_value_count}
\alias{unique_value_count}
\title{Calculates the number of unique values in a given vector}
\usage{
unique_value_count(value_vector, proportion = FALSE, remove_na = FALSE)
}
\arguments{
\item{value_vector}{A vector of values}

\item{proportion}{A logical indicating whether to return the proportion of unique values}

\item{remove_na}{A logical indicating whether to exclude NA values from the calculation}
}
\value{
The number of unique values in the input vector, or the proportion of unique values
if the 'proportion' argument is set to TRUE
}
\description{
This function calculates the number of unique values in a given vector.
The calculation can be adjusted to exclude NA values and return the proportion of unique values instead of the count.
}
