% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_class.R
\name{cr_class}
\alias{cr_class}
\title{Sub-classification of public clonotype}
\usage{
cr_class(
  group_count,
  margin = 1,
  min_freq = 1,
  public_min = 2,
  exclusive_max = 1
)
}
\arguments{
\item{group_count}{numerical vector / or multiple dimensional data-type (data frame / matrix / array) of number of colonotype in each group}

\item{margin}{the dimension to apply}

\item{min_freq}{minimal value per group}

\item{public_min}{minimum total sum of the}

\item{exclusive_max}{maximum number of groups that could be conciser exclusive}
}
\value{
character vector
}
\description{
computes the CR-class of each clonotype by its frequency in each group of repertoires.
}
\examples{

df_lists <- replicate(4, rand_group(), FALSE)
rand_df <- group_join(df_lists)
tbl <- share_table(rand_df, "clonotype", "group", "rep_id")
cr_class(head(tbl, 5))

}
