% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_class.R
\name{cr_class}
\alias{cr_class}
\title{Sub-classification of public clonotype}
\usage{
cr_class(share, public_min = 1L, exclusive_min = 1L)
}
\arguments{
\item{share_tbl}{}
}
\description{
Sub-classification of public clonotype
}
