% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{cr_class}
\alias{cr_class}
\title{Public clonotype sub-class of repertoire groups}
\usage{
cr_class(.data, .clonotype, .rep_id, .group_id, ...)
}
\arguments{
\item{.data}{repertoire data frame}

\item{.clonotype}{clonotype sequence variable}

\item{.rep_id}{repertoire identification column}

\item{.group_id}{group identification column}

\item{...}{additional variables to group by}
}
\value{
data frame with 'public' column
}
\description{
Public clonotype sub-class of repertoire groups
}
\examples{

df <- data.frame(
  group_id=gl(2,4,8,labels = c('cancer','control')),
  sample_id=gl(4,2,8,labels = LETTERS[1:4]),
  nt=c('CGCGTGAAG', 'CGGGTGAAG','CACGAA','AAGGGGTCCGTG',
       'AAGGGGTCCGTC','CGGGTGAAG','AAGGGGTCCGTT','CGGGTGAAG'),
  aa=c('RVK','RVK','HE','KGSV','KGSV','RVK','KGSV','RVK')
)

cr_class(df, aa, sample_id, group_id)

}
