% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level}
\alias{cr_level}
\title{CR level table}
\usage{
cr_level(df, clone, clonotype, rep_id, ...)
}
\arguments{
\item{...}{repertoire list / data frames vectors}
}
\value{
table
}
\description{
CR level table
}
\examples{

# generate random
df_lists <- replicate(4, rand_group(), FALSE)
rand_df <- group_join(df_lists)
names(rand_df)
unite(col = "id", c("rep_id" ,"group"), sep = "/")
id_df <- unite(rand_df, col = "id", c("group", "rep_id"), sep = "/")
cr_level(df = id_df, clone = "clone", clonotype = "clonotype", rep_id = "id")

}
