% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level}
\alias{cr_level}
\title{Count the number of unique clonal sequences in a nucleotide sequence.}
\usage{
cr_level(
  nucleotide,
  amino_acid,
  proportions = FALSE,
  ignore_na = FALSE,
  names = FALSE
)
}
\arguments{
\item{nucleotide}{a character vector containing the nucleotide sequence.}

\item{amino_acid}{a character vector containing the corresponding amino acid
sequence. If not provided, the nucleotide sequence will be automatically
translated to an amino acid sequence.}

\item{proportions}{a logical value indicating whether to return the proportions of
unique clonal sequences instead of their counts.}

\item{ignore_na}{a logical value indicating whether to remove missing values
before computing the result.}

\item{names}{a logical value indicating whether to return the result as a
named vector.}
}
\value{
a numeric vector containing the number of unique clonal sequences in
the nucleotide sequence. If \code{named} is \code{TRUE}, the result will be a named
vector.
}
\description{
This function takes a nucleotide sequence and a corresponding amino acid
sequence, and returns the number of unique clonal sequences in the nucleotide
sequence. The nucleotide sequence is automatically translated to an amino acid
sequence if necessary. The output can be returned as a named or unnamed vector.
}
\examples{
nucleotide <- c("ATG", "TGA", "TGA", "TAA", "TAA")
amino_acid <- c("M", "W", "W", "*", "*")
cr_level(nucleotide, amino_acid)
  # returns: 2

nucleotide <- c("ATG", "TGA", "TGA", "TAA", "TAA")
cr_level(nucleotide)
  # returns: 2

}
