% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level}
\alias{cr_level}
\title{Convergent recombination level (CR-level) of clonal sequences}
\usage{
cr_level(..., always_named = FALSE)
}
\arguments{
\item{...}{nucleotide sequences}

\item{always_named}{name single value. Default \code{TRUE}.}
}
\value{
CR-level integer vector (named by the clonotype)
}
\description{
calculate unique number of nucleotide sequences encoding a specific amino
acid sequence. This value associated with TCR repertoire clonal sequences
and termed as "CR-level" (convergent recombination level).
}
\examples{
cr_level("ATGTTT", "ATCTTC", "ATGTTC", "AAA")

}
