% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level}
\alias{cr_level}
\title{calculate the convergent recombination level (CR level) of a sample's clonotypes}
\usage{
cr_level(clonotype, match_clones_with)
}
\arguments{
\item{clonotype}{character vectors}

\item{match_clones_with}{same size as 'clonotype' character vectors}
}
\description{
Determine the convergent recombination level (CR level) by counting the number of times each clonotype occurs.
}
\examples{

# From the letter vector representing the clonotype sequences,
# generate a random list of letters.
clonotype_vec <- sample(LETTERS[1:10], 20, replace = TRUE)

clone_vec <- sample(LETTERS[1:2], 20, replace = TRUE)


}
