% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{cr_level}
\alias{cr_level}
\title{CR level of clonotyoes in a repertoire}
\usage{
cr_level(.data, .clone, ...)
}
\arguments{
\item{.data}{repertoire data frame}

\item{.clone}{clonal sequence variable}

\item{...}{additional variables to group by}
}
\value{
dataframe with \code{CRlevel} column
}
\description{
CR level of clonotyoes in a repertoire
}
\examples{

df <- data.frame(
group_id=gl(2,4,8,labels = c('cancer','control')),
sample_id=gl(4,2,8,labels = LETTERS[1:4]),
nt=c('CGCGTGAAG', 'CGGGTGAAG','CACGAA','AAGGGGTCCGTG',
     'AAGGGGTCCGTC','CGGGTGAAG','AAGGGGTCCGTT','CGGGTGAAG'),
aa=c('RVK','RVK','HE','KGSV','KGSV','RVK','KGSV','RVK')
)
cr_level(df, nt)
cr_level(df, nt, sample_id)

require('dplyr')

group_by(df, group_id) \%>\% cr_level(nt)
group_by(df, group_id) \%>\% cr_level(nt, sample_id)

}
