% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level}
\alias{cr_level}
\title{Convergent recombination level calculation.}
\usage{
cr_level(..., na.rm = FALSE)
}
\arguments{
\item{...}{Vectors (or list of vectors) of equal-length or data frame.}

\item{rm.na}{if TRUE, ignore NA.}
}
\value{
integer
}
\description{
unique number of values / rows.
}
\examples{

cr_level(c("A", "B","A", "B", "B", NA))

cr_level(c("s", "s", "a", "a"), c(1:3,NA), na.rm = FALSE)

}
