% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level}
\alias{cr_level}
\title{CR level table}
\usage{
cr_level(...)
}
\arguments{
\item{...}{repertoire list / data frames vectors}
}
\value{
table
}
\description{
CR level table
}
\examples{

g1 <- group_gen()
g2 <- group_gen()
g3 <- group_gen()
g4 <- group_gen()

# generate random
g <- list(g1, g2, g3, g4)

g <- lapply(seq_along(g), function(i) g[[i]] \%>\%
     setNames(paste0(LETTERS[i], seq_along(.))))

cr_level(g)



}
