% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level}
\alias{cr_level}
\title{Convergent recombination level calculation}
\usage{
cr_level(.clone, ...)
}
\arguments{
\item{.clone}{character vector of the clonal sequences.}

\item{...}{Additional vectors of clonal features to distinct same clonal sequences optional}
}
\value{
integer of unique count.
}
\description{
Convergent recombination level calculation
}
\examples{

cr_level(rep(LETTERS[1:2], each=4), gl(2,2,8), gl(4,3,8))

}
