% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level}
\alias{cr_level}
\title{Convergent Recombination level (CR-level)}
\usage{
cr_level(clonal_sequence, clonotype_sequence)
}
\arguments{
\item{clonal_sequence}{x}

\item{clonotype_sequence}{x}
}
\value{
data frame
}
\description{
Convergent Recombination level (CR-level)
}
\examples{

cr_level(c("ATG", "TTT", "ATC", "TTC", "TTG", "TTA"))

}
