% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level.R
\name{cr_level_df}
\alias{cr_level_df}
\title{Convergent Recombination level (CR-level)}
\usage{
cr_level_df(df, clonal_var, clonotype_var)
}
\arguments{
\item{df}{c}

\item{clonal_var}{c}

\item{clonotype_var}{x}
}
\value{
x
}
\description{
Convergent Recombination level (CR-level)
}
\examples{

clone_vec <- c("ATG", "TTT", "ATC", "TTC", "TTG", "TTA")

(df <- rep2DF(clone_vec))

cr_level_df(df, clone, clonotype)

clone_list <- list(c("ATG", "TTT", "ATC"), c("TTC", "TTG", "TTA"))

(df2 <- repList2DF(clone_list))

cr_level_df(dplyr::group_by(df2, rid))


}
