% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_level_matrix.R
\name{cr_level_matrix}
\alias{cr_level_matrix}
\title{claculate the CR level in each clonotype in every sample}
\usage{
cr_level_matrix(clonotype_list)
}
\arguments{
\item{clonotype_list}{}
}
\description{
Here, we define CR level as the number of different NT sequences encoding a specific AA sequence.
}
\examples{
# Generate a random list of letters (as the clonotype sequences) from the letters vector.
l <- replicate(6, sample(LETTERS, sample(13:18, 1, replace = TRUE)), simplify = FALSE)
cr_level_table(l) # number of vectors a letter found in


}
