% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crtcr-package.R
\docType{package}
\name{crtcr-package}
\alias{crtcr}
\alias{crtcr-package}
\title{crtcr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
