% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_class.R
\name{cr_class}
\alias{cr_class}
\title{CR-class of a clonotype by its share-level among groups.}
\usage{
cr_class(shared, min_shared = 1, min_public = 2, max_exclusive = 1)
}
\arguments{
\item{shared}{numerical vector}

\item{min_shared}{minimum numerical value}

\item{min_public}{minimum sum from all frequencies having a minimal value to be considered as public clonotype}

\item{max_exclusive}{maximal number of public clonotypes groups to be considered as public exclusive. o.w. inclusive.}
}
\value{
character object of "private" / "inclusive" / "exclusive" / NA_character_
}
\description{
CR-class of a clonotype by its share-level among groups.
}
\section{Functions}{
\itemize{
\item \code{cr_class()}: clonotype frequencies in each group, calculate its convergent
recombination public classifications based on various conditions.

}}
\examples{

cr_class(c(1,0,1,0)) # public inclusive

cr_class(c(2,0,0,0)) # public exclusive

cr_class(c(1,0,0,0)) # private

}
