% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_join.R
\name{group_join}
\alias{group_join}
\title{Join all repertoires' clonotypes data into one data frame}
\usage{
group_join(...)
}
\arguments{
\item{...}{repertoire group}
}
\value{
data frame
}
\description{
Join all repertoires' clonotypes data into one data frame
}
\examples{

nt_gen <- clone_gen()

group1 <- list(sapply(rpois(100, 2)+1, nt_gen), sapply(rpois(100, 2)+1, nt_gen))
group2 <- list(sapply(rpois(100, 2)+1, nt_gen), sapply(rpois(100, 2)+1, nt_gen))

group_join(list(a=group1, b=group2))


}
