% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusive.R
\name{is_exclusive}
\alias{is_exclusive}
\title{Is exclusive public clonotype}
\usage{
is_exclusive(.subgroup, .exclusive = 1, .is_public = NULL, na.rm = FALSE)
}
\arguments{
\item{.subgroup}{sub-group identification vector.}

\item{.exclusive}{maximal number of subgroups a clone is found in.}

\item{.is_public}{logical value of of wheather or not the clonotype is public. if FALSE, meaning it's private, return NA.}

\item{na.rm}{return \code{FALSE} instead of \code{NA} when \code{.is_public} is FALSE}

\item{...}{additional identifier vectors}
}
\value{
TRUE for exclusive, FALSE for inclusive.
}
\description{
True for if clonotype’s present across subgroups doesn't surpass the threshold
}
\examples{

is_exclusive(c("A", "A", "A", "A"))
is_exclusive(c("A", "A", "A", "A", "B"))

}
