% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_exclusive.R
\name{is_exclusive}
\alias{is_exclusive}
\title{logical check if a frequencies vector is considered public-exclusive with optional modifications}
\usage{
is_exclusive(shared, min_shared = 1L, min_public = 1L, max_exclusive = 1L)
}
\arguments{
\item{shared}{numerical vector}

\item{min_shared}{minimal value to be counted as repertoire having a clonotype}

\item{min_public}{minimal number of repertoire having a clonotype to be conciser public}

\item{max_exclusive}{upper limit to number of groups the could be considered exclusive}
}
\value{
TRUE for exclusive, FALSE for inclusive, and NA if non of the groups are public
}
\description{
logical check if a frequencies vector is considered public-exclusive with optional modifications
}
\examples{

is_exclusive(c(0,0,0,3)) # TRUE (exclusive public)

is_exclusive(c(0,1,2,3)) # FALSE (inclusive public)

is_exclusive(c(0,1,2,3), min_shared=3) # TRUE (exclusive public)

is_exclusive(c(0,1,2,3), min_shared=4) # NA (not public)

is_exclusive(c(0,1,2,3), min_public=7) # NA (not public)

is_exclusive(c(0,1,2,3), max_exclusive=3) # TRUE (exclusive public - up to three groups)

is_exclusive(c(0,1,2,3), max_exclusive=2) # FALSE (inclusive public - more than two groups)

is_exclusive(c(0,1,2,3), min_shared=2, max_exclusive=2) # TRUE (exclusive public - up to two groups)

}
