% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_exclusive.R
\name{is_exclusive}
\alias{is_exclusive}
\title{logical check if a frequencies vector is considered public-exclusive with optional modifications}
\usage{
is_exclusive(gruop_freq, max_exclusive = 1)
}
\arguments{
\item{max_exclusive}{upper limit to number of groups the could be considered exclusive}

\item{shared}{numerical vector}
}
\value{
TRUE for exclusive, FALSE for inclusive, and NA if non of the groups are public
}
\description{
logical check if a frequencies vector is considered public-exclusive with optional modifications
}
\examples{

is_exclusive(c(0,0,0,3)) # TRUE (exclusive public)

}
