% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_class.R
\name{is_exclusive}
\alias{is_exclusive}
\title{logical check if a frequencies vector is considered public-exclusive with optional modifications}
\usage{
is_exclusive(freq, id, min_freq = 1L, min_rep = 1L, max_group = 1L)
}
\arguments{
\item{freq}{freq >= min_freq}

\item{id}{same length vector to split the groups by}

\item{min_freq}{minimal value to be counted as repertoire having a clonotype}

\item{min_rep}{minimal number of repertoire having a clonotype to be conciser public}

\item{max_group}{upper limit to number of groups the could be considered exclusive}
}
\value{
TRUE for exclusive, FALSE for inclusive, and NA if non of the groups are public
}
\description{
logical check if a frequencies vector is considered public-exclusive with optional modifications
}
\examples{

is_exclusive(c(1, 0 , 2, 4), gl(2,2), min_freq=2)

}
