% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_exclusive.R
\name{is_exclusive}
\alias{is_exclusive}
\title{Title x}
\usage{
is_exclusive(..., .exclusive = 1)
}
\arguments{
\item{...}{x}

\item{.exclusive}{x}

\item{na.rm}{x}
}
\value{
x
}
\description{
Title x
}
\examples{


require(dplyr)

df <- data.frame(x=factor(LETTERS[sample(1:5, 5,replace = TRUE)]),
                 y=factor(letters[sample(1:2, 5, replace = TRUE)]),
                 clonotype=sample(c("ASD", "AWD"), 5, replace = TRUE, prob = c(.2,.8)))

df \%>\%
  group_by(clonotype) \%>\%
  summarise(
    public=is_public(x,y),
    exclusive=is_exclusive(y)
  )



}
