% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_exclusive_freq.R
\name{is_exclusive_freq}
\alias{is_exclusive_freq}
\title{Is exclusive public clonotype by frequencies}
\usage{
is_exclusive_freq(.freq, .subgroup, .exclusive = 1, na.rm = FALSE)
}
\arguments{
\item{.freq}{numerical vector}

\item{.exclusive}{value count threshold}

\item{na.rm}{if FALSE (the default), return NA when no frequency value meet with the condition}
}
\value{
TRUE for exclusive, FALSE for inclusive.
}
\description{
Is exclusive public clonotype by frequencies
}
\examples{


is_exclusive_freq(c(1,1,0,0,0), c("A", "A", "B", "B", "C"))
is_exclusive_freq(c(1,1,0,0,1), c("A", "A", "B", "B", "C"))

is_exclusive_freq(c(1,1,0,0,1), c("A", "A", "B", "B", "C"), 2)
is_exclusive_freq(c(1,1,0,0,0), c("A", "A", "B", "B", "C"), 1/3)


}
