% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_class.R
\name{is_public}
\alias{is_public}
\title{logical check if a frequencies vector is considered public with optional modifications}
\usage{
is_public(freq, min_freq = 1L, public_min = 2L, share_level = TRUE)
}
\arguments{
\item{freq}{numerical vector}

\item{min_freq}{minimal value to be counted as repertoire having a clonotype}

\item{public_min}{minimal number of repertoire having a clonotype to be conciser public}

\item{share_level}{TRUE if frequencies represent share-level (number of repertoires having a clonotype in each group), and FALSE for CR-level representation (number of same-clonotype clone in a repertoire)}
}
\value{
TRUE if its public, FALSE if private, and NA if none of the frequencies are sufficient
}
\description{
logical check if a frequencies vector is considered public with optional modifications
}
\examples{

is_public(c(1, 0 , 2, 4), min_freq=3)

}
