% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_public.R
\name{is_public}
\alias{is_public}
\title{Public clonotype frequencies}
\usage{
is_public(freq, .public = 2)
}
\arguments{
\item{freq}{Numerical vector.}

\item{.public}{Minimum number of repertoires}
}
\value{
logical scalar. TRUE for public, FALSE for private (or none)
}
\description{
check if clonotype's frequencies throughout repertoire group are public
}
\examples{

is_public(c(0,1,0,3))

}
