% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_public.R
\name{is_public}
\alias{is_public}
\title{is clonotype public by sub-groups share-level}
\usage{
is_public(shared, min_shared = 1, min_public = 2 * min_shared)
}
\arguments{
\item{shared}{numerical vector}

\item{min_shared}{minimal value to be counted as repertoire having a clonotype}

\item{min_public}{minimal number of repertoire having a clonotype to be conciser public}
}
\value{
TRUE if its public, FALSE if private, and NA if none of the frequencies are sufficient
}
\description{
is clonotype public by sub-groups share-level
}
\examples{

is_public(c(0,1,0,0)) # FALSE (private clonotype)

is_public(c(0,2,0,0)) # FALSE (public clonotype)

is_public(c(0,2,0,0), min_public = 3) # FALSE (private clonotype)

is_public(c(0,1,2,3), min_public = 5) # TRUE (public clonotype)

is_public(c(0,2,0,0), min_shared=3) # NA (value is not sufficient)

is_public(c(0,1,2,3), min_shared=3) # TRUE (public clonotype)

is_public(c(0,1,2,3), min_shared=3, min_public = 5) # FALSE (private clonotype)

}
