% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_public.R
\name{is_public}
\alias{is_public}
\title{Public clonotype frequencies}
\usage{
is_public(..., .public = 2)
}
\arguments{
\item{...}{x}

\item{.public}{Minimum number of repertoires}

\item{na.rm}{x}
}
\value{
logical scalar. TRUE for public, FALSE for private (or none)
}
\description{
check if clonotype's frequencies throughout repertoire group are public
}
\examples{

is_public(c("A", "B", "A"))

require(dplyr)

df <- data.frame(x=factor(LETTERS[sample(1:5, 10,replace = TRUE)]),
                 y=factor(letters[sample(1:3, 10, replace = TRUE)]),
                 clonotype=sample(c("ASD", "AWD"), 10, replace = TRUE))

df \%>\%
  group_by(clonotype) \%>\%
  summarise(public=is_public(x,y, .public = .5))

}
