% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_public.R
\name{is_public}
\alias{is_public}
\title{Is public clonotype}
\usage{
is_public(.id, .public = 2)
}
\arguments{
\item{.id}{repertoire identifier vector}

\item{.public}{minimum number of repertoires.}

\item{...}{additional identifier vectors}
}
\value{
TRUE for public, FALSE for private.
}
\description{
Is public clonotype
}
\examples{

is_public(c("A", "A", "A", "A"))
is_public(c("A", "A", "A", "A", "B"))

}
