% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_class.R
\name{public_class}
\alias{public_class}
\title{Public classification}
\usage{
public_class(.id, .subgroup, ...)
}
\arguments{
\item{.id}{repertoire identifier vector}

\item{.subgroup}{sub-group identification vector.}

\item{...}{additional identifier vectors.}
}
\value{
"private" / "exclusive” / “inclusive”
}
\description{
returns the publicness classifications "private," "exclusive,” and
“inclusive” by the repertoires’ presence, throughout all subgroups.
}
\examples{

public_class(c("A", "A"), c("C", "D"))
public_class(c("A", "B"), c("C", "C"))
public_class(c("A", "A"), c("C", "C"))

}
