% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{public_class}
\alias{public_class}
\title{Public clonotype in repertoire collection}
\usage{
public_class(.data, .clonotype, .rep_id, ...)
}
\arguments{
\item{.data}{repertoire data frame}

\item{.clonotype}{clonotype sequence variable}

\item{.rep_id}{repertoire identification column}

\item{...}{additional variables to group by}
}
\value{
data frame with 'public' column
}
\description{
Public clonotype in repertoire collection
}
\examples{

df <- data.frame(
  group_id=gl(2,4,8,labels = c('cancer','control')),
  sample_id=gl(4,2,8,labels = LETTERS[1:4]),
  nt=c('CGCGTGAAG', 'CGGGTGAAG','CACGAA','AAGGGGTCCGTG',
       'AAGGGGTCCGTC','CGGGTGAAG','AAGGGGTCCGTT','CGGGTGAAG'),
  aa=c('RVK','RVK','HE','KGSV','KGSV','RVK','KGSV','RVK')
)

public_class(df, aa, sample_id, group_id)

}
