% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_prepare.R
\name{repList2DF}
\alias{repList2DF}
\title{list of repertoire to data frame}
\usage{
repList2DF(...)
}
\arguments{
\item{...}{for data frames with more than one column, clonal sequence column name is needed. adding also clonotype column name is suggested}

\item{rep_list}{list of character vectors or data frames}
}
\value{
data frame
}
\description{
list of repertoire to data frame
}
\examples{

vec_list <- list(
A=c("ATG", "TTC", "TTT"),
B=c("TTT", "ATG", "ATG")
)
vec_list

repList2DF(vec_list)

df_list <- list(
  A=data.frame(nt=c("ATG", "TTC", "TTT"),
               aa=c("M", "F", "F")),
  B=data.frame(nt=c("TTT", "ATG", "ATG"),
               aa=c("F", "M", "M"))
)

df_list

repList2DF(df_list, nt)
repList2DF(df_list, nt, aa)

}
