% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_level.R
\name{share_level}
\alias{share_level}
\title{Share-level of the group's repertoires}
\usage{
share_level(...)
}
\arguments{
\item{...}{}
}
\value{
share-level vector named by clonotype sequence
}
\description{
Share-level of the group's repertoires
}
\examples{

rep1 <- c('CGGGTGAAG', 'CGGGTG', 'CGGGTGAAG','CACGAA','AAGGGGTCCGTG')
rep2 <- c('AAGGGGTCCGTC','CGGGTGAAG','AAAGGGTCCGTT','CGGGTCAAG')
rep3 <- c('AAGGGGTCAGTC','CGTGTGAAG','AAGGGGTCCGTT','CGGGTCAAG')

share_level(rep1, rep2, rep3)

share_level(c(rep1, rep2), c(rep2, rep3))


}
