% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_level.R
\name{share_level}
\alias{share_level}
\title{Compute the share level of clonotypes in a group}
\usage{
share_level(data, rid, clonotype_var)
}
\arguments{
\item{data}{data frame}

\item{rid}{repertoire identification column name}

\item{clonotype_var}{clonotype sequence column name}
}
\value{
dataframe
}
\description{
Compute the share level of clonotypes in a group
}
\examples{

require(dplyr)

rand_group() \%>\%
groupList2DF() \%>\%
group_by(gid) \%>\%
share_level(rid, clonotype)

}
