% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharing_level.R
\name{share_level}
\alias{share_level}
\title{count clonotype-sharing among individuals}
\usage{
share_level(clonotype_list)
}
\arguments{
\item{clonotype_list}{list of character vectors}
}
\value{
data frame - number of vectors each sequence presented in
}
\description{
for a given list of amino acid sequences vectors of each individual/sample count number of samples each clonotype is presented
}
\examples{
# Generate a random list of letters (as the clonotype sequences) from the letters vector.
l <- replicate(6, sample(LETTERS, sample(13:18, 1, replace = TRUE)), simplify = FALSE)
share_level(l) # number of vectors a letter found in

}
