% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_level.R
\name{share_level}
\alias{share_level}
\title{Share-level of the group's repertoires}
\usage{
share_level(rep_id, clonotype)
}
\arguments{
\item{rep_id}{repertoire ID vector}

\item{clonotype}{clonotype sequences vector}
}
\value{
share-level vector named by clonotype sequence
}
\description{
Share-level of the group's repertoires
}
\examples{

sample_id <- rep(c("A", "B"), each=3)

clonotype_vec <- c("M", "F", "I", "F", "L", "L")

share_level(sample_id, clonotype_vec)

}
