% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_level.R
\name{share_table}
\alias{share_table}
\title{Title}
\usage{
share_table(group_list, share_method = "unique")
}
\arguments{
\item{group_list}{x}

\item{share_method}{x}
}
\value{
x
}
\description{
x
}
\examples{

gl <- replicate(4, rand_group())

share_table(gl, share_method = "CRlevel")


}
