% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharing.R
\name{sharing}
\alias{sharing}
\title{Title}
\usage{
sharing(...)
}
\arguments{
\item{...}{data frames or vectors}

\item{by}{x}
}
\value{
table
}
\description{
Title
}
\examples{

nt_gen <- clone_gen()

group1 <- list(sapply(rpois(, 5)+1, nt_gen), sapply(rpois(,5)+1, nt_gen))
group2 <- list(sapply(rpois(100, 7)+1, nt_gen), sapply(rpois(100, 8)+1, nt_gen))
group3 <- list(sapply(rpois(100, 2)+1, nt_gen), sapply(rpois(100, 2)+1, nt_gen))
group4 <- list(sapply(rpois(100, 2)+1, nt_gen), sapply(rpois(100, 2)+1, nt_gen))
group5 <- list(sapply(rpois(100, 2)+1, nt_gen), sapply(rpois(100, 2)+1, nt_gen))

groups <- group_join(list(a=group1, b=group2, c=group3, d=group4, e=group5))

}
