% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharing.R
\name{sharing}
\alias{sharing}
\title{Title}
\usage{
sharing(...)
}
\arguments{
\item{...}{data frames or vectors}
}
\value{
table
}
\description{
Title
}
\examples{

# generate random
g <- replicate(4, rand_group(), simplify = FALSE)
sharing(g)

}
