% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Converts nucleotides to amino acids}
\usage{
translate(nt_vec)
}
\arguments{
\item{nt_vec}{DNA sequences as character vector composed with the letters
"A", "G", "T" or"C" (case-sensitive to upper case).}
}
\value{
Character vector of the coding amino acid sequences.
}
\description{
Converts a character vector of nucleotide sequences (uppercase 'A','G','T','C')
to their corresponding amino acid sequences by decoding consecutive triplets
with the codon table.
}
\examples{

translate(c("AGTTTA", "ATGCCT", "GGTTGAAAA")) # asterisk ('*') is a stop codon

}
