% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Converts nucleotide sequence to amino acid sequence}
\usage{
translate(nt_vec)
}
\arguments{
\item{nt_vec}{DNA sequences as character vector composed with the letters
"A", "G", "T" or"C" (case-sensitive to upper case).}
}
\value{
Character vector of the coding amino acid sequences.
}
\description{
Converts nucleotide sequence to amino acid sequence
}
\examples{

translate(c("AGTTTA", "ATGCCT", "GGTTGAAAA")) # asterisk ('*') is a stop codon

}
