% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{translate}
\alias{translate}
\title{Translates a DNA sequence into an amino acid sequence.}
\usage{
translate(nt_vec)
}
\arguments{
\item{nt_vec}{}
}
\description{
Translates a DNA sequence into an amino acid sequence.
}
