% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nt2aa.R
\name{translate}
\alias{translate}
\title{Converts nucleotide sequence to amino acid sequence}
\usage{
translate(nt_vec)
}
\arguments{
\item{nt_vec}{character vector with the letters "A", "G", "T", "C" (lower/upper case)}
}
\value{
character vector of the coding amino acid sequences.
}
\description{
Converts nucleotide sequence to amino acid sequence
}
\examples{

translate(c("AGTTTA", "ATGCCT", "GGTTGAAAA")) # asterisk ('*') is a stop codon

}
